+++ packages/falso/src/index.ts
@@ -1,3 +1,4 @@
+export { seed } from './lib/random';
 export { zipCode } from './lib/zip-code';
 export { zipCodeByState } from './lib/zip-code-by-state';
 export { city } from './lib/city';