+++ packages/falso/src/lib/random-image.ts
@@ -1,19 +1,6 @@
 import { rand } from './core';
+import { data } from './random-image.json';
 
 export function randomImage() {
-  return rand([
-    'http://placeimg.com/640/480/food',
-    'http://placeimg.com/640/480/animals',
-    'http://placeimg.com/640/480/nightlife',
-    'http://placeimg.com/640/480/cats',
-    'http://placeimg.com/640/480/people',
-    'http://placeimg.com/640/480/transport',
-    'http://placeimg.com/640/480/abstract',
-    'http://placeimg.com/640/480/technics',
-    'http://placeimg.com/640/480/nature',
-    'http://placeimg.com/640/480/city',
-    'http://placeimg.com/640/480/business',
-    'http://placeimg.com/640/480/sports',
-    'http://placeimg.com/640/480/fashion',
-  ]);
+  return rand(data);
 }