+++ packages/falso/src/lib/random.ts
@@ -1,0 +1,11 @@
+import seedRandom from 'seedrandom';
+
+let prng = seedRandom();
+
+export function random() {
+  return prng();
+}
+
+export function seed(value?: string) {
+  prng = seedRandom(value);
+}