+++ packages/falso/src/lib/transaction-type.ts
@@ -1,5 +1,6 @@
 import { rand } from './core';
+import { data } from './transaction-type.json';
 
 export function transactionType() {
-  return rand(['deposit', 'withdrawal', 'payment', 'invoice']);
+  return rand(data);
 }