+++ packages/falso/CHANGELOG.md
@@ -2,18 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-# [1.2.0](https://github.com/ngneat/falso/compare/falso-1.1.0...falso-1.2.0) (2022-01-10)
+# [1.3.0](https://github.com/ngneat/falso/compare/falso-1.2.0...falso-1.3.0) (2022-01-11)
 
+### Bug Fixes
+
+- 🐞 produce cjs import ([5af2dc1](https://github.com/ngneat/falso/commit/5af2dc1b2330fb0313c5d395e09f3f77b6e39e89)), closes [#16](https://github.com/ngneat/falso/issues/16)
+- move fixed version seedrandom to falso ([fc28435](https://github.com/ngneat/falso/commit/fc28435af2bc934f61e64dcfdca26de513437ad4))
+- move types to dev deps, adjust gitignore ([50033ff](https://github.com/ngneat/falso/commit/50033ff1d7c750ec6c8b41dd9ad5a45ecf949297))
 
 ### Features
 
-* 🔥 update rgb() to return dynamically generated rgb string ([d4c1cf7](https://github.com/ngneat/falso/commit/d4c1cf795974835c77971f2947211666e682487d))
+- add seed function ([b876f51](https://github.com/ngneat/falso/commit/b876f51720fa9c7f1a5d9aaf15739fd06d87f37a)), closes [#8](https://github.com/ngneat/falso/issues/8)
 
+# [1.2.0](https://github.com/ngneat/falso/compare/falso-1.1.0...falso-1.2.0) (2022-01-10)
 
+### Features
+
+- 🔥 update rgb() to return dynamically generated rgb string ([d4c1cf7](https://github.com/ngneat/falso/commit/d4c1cf795974835c77971f2947211666e682487d))
 
 # [1.1.0](https://github.com/ngneat/falso/compare/falso-1.0.0...falso-1.1.0) (2022-01-10)
 
-
 ### Features
 
-* 🔥 add rgb function ([f356fee](https://github.com/ngneat/falso/commit/f356feeba6c53a82dd22260dd14dfb762680e0dc))
+- 🔥 add rgb function ([f356fee](https://github.com/ngneat/falso/commit/f356feeba6c53a82dd22260dd14dfb762680e0dc))